requirements.txt
selenium
chromedriver_autoinstaller
python-dotenv
undetected-chromedriver
